requirements.txt

# 🔧 Core Django
django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt>=5.3.1
asgiref==3.9.1
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2
dj-database-url==3.0.1
psycopg[binary]>=3.2.1

# 🚀 WSGI Server
gunicorn==23.0.0

# 🎨 Static Files
whitenoise==6.10.0

# 🔄 Background Tasks
celery==5.5.3
django-celery-beat==2.8.1
django-timezone-field==7.1
redis==6.4.0
amqp==5.3.1
billiard==4.2.1
kombu==5.5.4
vine==5.1.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==2.0.6
python-crontab==3.3.0

# 🤖 Machine Learning (CPU-only)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu
torchvision==0.23.0+cpu
scikit-learn==1.7.1
imbalanced-learn==0.14.0
numpy==2.2.6
scipy==1.16.1
pandas==2.3.2
joblib==1.5.2
threadpoolctl==3.6.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.5

# 🖼️ Image Processing
opencv-python-headless==4.12.0.88
pillow==11.3.0

# 🗺️ GIS & Geospatial
geopandas==1.1.1
pyogrio==0.11.1
pyproj==3.7.2
shapely==2.1.1

# 🌐 HTTP & Networking
requests[socks]==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pysocks==1.7.1

# 💰 Payment
mercadopago==2.3.0

# 🛠️ Utilities
python-dotenv==1.1.1
dotenv==0.9.9
python-dateutil==2.9.0.post0
six==1.17.0
packaging==25.0
filelock==3.19.1
fsspec==2025.7.0
typing-extensions==4.14.1

# 📄 HTML Parsing (se necessário)
beautifulsoup4==4.13.5
soupsieve==2.8

# 🎨 Template Engine
jinja2==3.1.6
markupsafe==3.0.2
gdown==5.2.0